refactor(UIX): :recycle: centralize LocalButton helper
